move exclude code to altwcs
