+++ stwcs/updatewcs/utils.py
@@ -6,12 +6,6 @@ import logging
 logger = logging.getLogger("stwcs.updatewcs.utils")
 
 
-# These are keywords that may be present in headers created by WCSLIB.
-# However, HST keeps them (or similar information) in the primary header and so
-# they should not be managed by STWCS when copying/archiving an HST WCS:
-EXCLUDE_WCSLIB_KWDS = ['EQUINOX', 'LONPOLE', 'LATPOLE', 'RESTWAV', 'RESTFRQ']
-
-
 def diff_angles(a, b):
     """
     Perform angle subtraction a-b taking into account
@@ -267,31 +261,3 @@ def remove_distortion(fname, dist_keyword):
     ext_mapping.sort()
     for hdu in ext_mapping[::-1]:
         del fname[hdu]
-
-
-def exclude_hst_specific(hdr, wcskey=' '):
-    """ Remove HST-specific keywords from header ``hdr`` that are not supposed
-    to be present in SCI headers.
-
-    """
-    if (wcskey is not None and
-        ((not isinstance(wcskey, str) or len(wcskey) != 1 or
-          wcskey.strip() not in string.ascii_uppercase))):
-        raise ValueError(
-            "Parameter 'wcskey' must be a character - one of 'A'-'Z' or ' '."
-        )
-
-    if 'EXTNAME' in hdr and hdr['EXTNAME'] not in ['SCI', 'DQ', 'ERR']:
-        logger.warning("Input header must be either 'SCI', 'DQ', or 'ERR' image headers.")
-        logger.warning("HST-specific keywords will not be excluded from the header.")
-        return hdr
-
-    if wcskey is None or wcskey == ' ':
-        wcskey = ''
-
-    for kwd in EXCLUDE_WCSLIB_KWDS:
-        kwda = kwd + wcskey
-        if kwda in hdr:
-            del hdr[kwda]
-
-    return hdr
